I've created a component to handle displaying properties within the sales and lettings pages depending on whether a filter has been applied or not.
