+++ root/frontend/src/components/FilteredProperties/FilteredProperties.js
@@ -2,8 +2,7 @@ import React, { Fragment, useState, useEffect } from 'react';
 import { fetchProperties } from '../../tools/functions';
 import { ADD_PROPERTIES } from '../../actions/types';
 import './FilteredProperties.scss';
-import currencyFormatter from '../../tools/currencyFormatter';
-import Spinner from '../Spinner/Spinner';
+import DisplayProperties from '../DisplayProperties/DisplayProperties';
 import store from '../../store';
 
 export default function FilteredProperties (props) {
@@ -33,25 +32,11 @@ export default function FilteredProperties (props) {
         <Fragment>
             <div className="grid r__container clear_search_form">
                 {
-                    properties.length > 0  ? properties.map((property, index) => {
-                        const price = currencyFormatter.format(property.price);
-                        if ( filter.length > 0 ) {
-
-                        }
-                        else if (property.contractType === contractType) return ( 
-                            <a href="#" key={ index } className="grid__item" style={{ backgroundImage: `url(data:image/png;base64,${property.image})` }}>
-                                <div className="grid__item-content">
-                                    <div className="grid__item-content__container">
-                                        <h4 className="title">{ property.location }</h4>
-                                        <h5 className="price">Price: { price }</h5>
-                                        <i className="fa fa-bed icon" aria-hidden="true"> { property.bedroom }</i>
-                                    </div>
-                                </div>
-                            </a> 
-                        )
-                    })
-                    :
-                    <Spinner />
+                    Object.keys(filter).length === 0 ?
+                        
+                        <DisplayProperties properties={properties} contractType = {contractType} />
+                        :
+                        <DisplayProperties properties = {properties} contractType = {contractType} filter = {filter} />
                 }
             </div>
         </Fragment>
